Fix codacy issue: package name
