+++ src/main/java/no/sikt/nva/pubchannels/channelregistry/model/search/ChannelRegistryEntityPageInformation.java
@@ -1,4 +1,4 @@
-package no.sikt.nva.pubchannels.channelRegistry.model.search;
+package no.sikt.nva.pubchannels.channelregistry.model.search;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import no.sikt.nva.pubchannels.handler.search.ThirdPartyPageInformation;
