+++ src/main/java/no/sikt/nva/pubchannels/channelregistry/model/search/ChannelRegistrySearchPublisherResponse.java
@@ -1,8 +1,8 @@
-package no.sikt.nva.pubchannels.channelRegistry.model.search;
+package no.sikt.nva.pubchannels.channelregistry.model.search;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import no.sikt.nva.pubchannels.channelRegistry.model.ChannelRegistryPublisher;
+import no.sikt.nva.pubchannels.channelregistry.model.ChannelRegistryPublisher;
 import no.sikt.nva.pubchannels.handler.search.ThirdPartySearchResponse;
 
 @JsonSerialize
