+++ src/main/java/no/sikt/nva/pubchannels/channelregistry/model/search/ChannelRegistrySearchSeriesResponse.java
@@ -1,12 +1,14 @@
-package no.sikt.nva.pubchannels.channelRegistry.model.search;
+package no.sikt.nva.pubchannels.channelregistry.model.search;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import no.sikt.nva.pubchannels.channelRegistry.model.ChannelRegistryJournal;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import no.sikt.nva.pubchannels.channelregistry.model.ChannelRegistrySeries;
 import no.sikt.nva.pubchannels.handler.search.ThirdPartySearchResponse;
 
-public record ChannelRegistrySearchJournalResponse(
+@JsonSerialize
+public record ChannelRegistrySearchSeriesResponse(
     @JsonProperty(ENTITY_PAGE_INFORMATION) ChannelRegistryEntityPageInformation pageInformation,
-    @JsonProperty(ENTITY_RESULT_SET) ChannelRegistryEntityResultSet<ChannelRegistryJournal> resultSet)
+    @JsonProperty(ENTITY_RESULT_SET) ChannelRegistryEntityResultSet<ChannelRegistrySeries> resultSet)
     implements ThirdPartySearchResponse {
 
     private static final String ENTITY_RESULT_SET = "entityResultSetDto";
