+++ src/test/java/no/sikt/nva/pubchannels/handler/TestUtils.java
@@ -38,10 +38,10 @@ import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import no.sikt.nva.pubchannels.channelRegistry.ChannelRegistryClient;
-import no.sikt.nva.pubchannels.channelRegistry.mapper.ScientificValueMapper;
-import no.sikt.nva.pubchannels.channelRegistry.model.ChannelRegistryLevel;
-import no.sikt.nva.pubchannels.channelRegistry.model.search.ChannelRegistryEntityPageInformation;
+import no.sikt.nva.pubchannels.channelregistry.ChannelRegistryClient;
+import no.sikt.nva.pubchannels.channelregistry.mapper.ScientificValueMapper;
+import no.sikt.nva.pubchannels.channelregistry.model.ChannelRegistryLevel;
+import no.sikt.nva.pubchannels.channelregistry.model.search.ChannelRegistryEntityPageInformation;
 import no.unit.nva.testutils.HandlerRequestBuilder;
 import nva.commons.core.SingletonCollector;
 import nva.commons.core.paths.UriWrapper;
