+++ src/test/java/no/sikt/nva/pubchannels/handler/fetch/journal/PublicationChannelMockClient.java
@@ -13,8 +13,8 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
-import no.sikt.nva.pubchannels.channelRegistry.mapper.ScientificValueMapper;
-import no.sikt.nva.pubchannels.channelRegistry.model.ChannelRegistryLevel;
+import no.sikt.nva.pubchannels.channelregistry.mapper.ScientificValueMapper;
+import no.sikt.nva.pubchannels.channelregistry.model.ChannelRegistryLevel;
 import no.sikt.nva.pubchannels.handler.ChannelRegistryBodyBuilder;
 import no.sikt.nva.pubchannels.handler.ScientificValue;
 import no.sikt.nva.pubchannels.handler.ThirdPartyJournal;
